direct deref - experiment
The idea of this commit is to change the 'deref' operation to match
against the available constructors on the ptr interacting with it,
allowing us to allocate only the needed parts of the reference - or, in
case of tail calls, only the final result. For example, consider:

@loop = (? (#0 (x y)) a a) & @loop ~ (x y)

It would be expanded / dereferenced as follows:

//@loop ~ (#100 R)
//-------------------------------------------
//apply @loop | (#100 R) ~ (? (#0 (x y)) a a)
//apply @loop | @loop    ~ (x y)
//-------------------------------------------
//apply @loop | #100  ~ ? (#0 (x y)) a
//apply @loop | R     ~ a
//apply @loop | @loop ~ (x y)
//-------------------------------------------
//apply @loop | *     ~ #0
//apply @loop | #99   ~ x
//apply @loop | a     ~ y
//apply @loop | R     ~ a
//apply @loop | @loop ~ (x y)
//-------------------------------------------
//apply @loop | #99   ~ x
//apply @loop | a     ~ y
//apply @loop | R     ~ a
//apply @loop | @loop ~ (x y)
//-------------------------------------------
//apply @loop | @loop ~ (#99 R)

Notice that, in the last like, we're back to where we started, except
with a smaller argument. This would go on continuously, without
performing any allocation, until we reach the ultimate result. Moreover,
this could be performed in a fast loop. I believe such approach could
wield significant speedups against the current implementation.

Sadly, this commit isn't complete yet and I have to stop working on it
due to my insomnia issues.