direct deref prototype: +89% perf on loops
this implement a new method for dereferencing that, instead of writing
the dereferenced net to memory, will attempt to match it with the other
side of the wire, performing interactions like con-con and num-mat
without allocating intermediate nodes. this allows tail-recursive
functions (like @lp below) to be performed in a loop, with constant
memory footprint. this achieves +89% perf on the following program:

@lpc  = (a b) & @lp ~ (a b)
@lp   = (? (#0 @lpc) a a)
@main = R & @lp ~ (#40000000 R)

this also paves the way to compiled HVMC programs, and it shows that the
current design of HVM is sufficient for compilation, and further
additions aren't necessary.

sadly, this new approach seems to be detrimental to programs that are
not loops, since it relies on a slower recursive allocator that doesn't
perform allocation directly in a memcpy-like fashion, as happened on the
old allocator. it also is still buggy, as some programs don't reduce
correctly. further work would be necessary to merge this. for the
moment, though, I don't think the added complexity justifies moving the
repo in that direction yet, but it provides important insights and
inspirations that may be important in the future.